okay, for real this time
Restarted, this time from the output system, since that was the most difficult to figure out.
Currently using an immediate-ish UI API, except this is just for output (all the input is done as a stream of events).
Gonna work on some concept art on top of the UI once I have it working, that should be fun!